Complete main and improve parsing
The parsing routines are now called on `main`. Also, only spaces and
tabs are now considered whitespace, since that allows the new line to
terminate the program (instead of waiting for EOD).

Parsing routine `deal_with_declarator` had the call to
`deal_with_pointers` moved to inside the stack loop. This differs from
the original algorithm, however it is necessary here since pointers do
not have special treatment in `get_token` (as in the original). Thus,
without repeatedly analysing for them in the stack, they would be
printed as `*` only.